Move -lowerswitch option to generate step.
Actually, original code has a bug.
It should be form of
"opt A -o B &&mv B A"
not "opt A -o A"
However, this kind of code didn't work
(fail on mv, there is no such file)
It might be problem of Parallel, so I separated
lower_switch so that lowering bundle of triples and then validate,
but it also failed.